functions/requirements.txt
fastapi
firebase-admin
firebase-functions
a2wsgi
python-dotenv
pydantic
httpx
sendgrid
uvicorn